+++ src/idtrackerai/utils/confparams.py
@@ -5,6 +5,10 @@ from dataclasses import asdict, dataclass
 class ConfParams:
     """Dataclass containing all CNN hyper-parameters. These can be modified by"""
 
+    DEVICE: str = ""
+    """This is just the user specified name of the device.
+    Use idtrackerai.base.network.DEVICE for a proper device object instead"""
+
     MODEL_AREA_SD_TOLERANCE: float = 4
     MINIMUM_NUMBER_OF_CROSSINGS_TO_TRAIN_CROSSING_DETECTOR: int = 10