+++ docs/source/user_guide/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 
 Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 
+- Increase and decrease GUI font size with ``Ctrl++`` and ``Ctrl+-``.
+
 5.2.11
 ======