+++ docs/source/user_guide/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 
 - Increase and decrease GUI font size with ``Ctrl++`` and ``Ctrl+-``.
+- Fix https://gitlab.com/polavieja_lab/idtrackerai/-/issues/90
 
 5.2.11
 ======