+++ docs/source/user_guide/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 
+- Works in Python 3.12, 3.11 and 3.10
 - Increase and decrease GUI font size with ``Ctrl++`` and ``Ctrl+-``.
 - Fix https://gitlab.com/polavieja_lab/idtrackerai/-/issues/90
 - Optimized `"Connecting coexisting fragments"`.