+++ docs/source/user_guide/changelog.rst
@@ -9,6 +9,7 @@ Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 - Fix https://gitlab.com/polavieja_lab/idtrackerai/-/issues/90
 - Optimized `"Connecting coexisting fragments"`.
 - The log file copy in session folder contains error tracebacks.
+- Add ``DEVICE`` as an optional input parameter.
 
 5.2.11
 ======